feature/280-progress-berechnen (#58)
* get keyresults and measures of only specific objective with database query

* use progressservice in controllers to calculate progress of objective

* add tests in ProgressServiceTest and fix failing tests

* move updating of progresservice under updating of measure

* use query to calculate progress and rewrite tests

* comment integrationtests because they don't work on github

* return 0 if progress is null directly in query

* disable tests

* calculate progress in one method and fix tests because of this change

* disable tests again because they don't work on github

* work on rebase

* fix tests after merging

* format code with backend-formatter

* delete unused variable in test

* implement progresservice in measurecontroller and change progress to long

* resolve requested changes in test

* disable progress tests

* correct query in ObjectiveRepository and add test to verify it

* optimize query with option to have progress calculation of objective without keyresults

* take basis value of keyresult if keyresult has no measures

* don't use keyresult with no measure in calculation

* format ObjectiveRepository

* Reenable ProgressServiceTestIT Tests

+++ backend/src/main/java/ch/puzzle/okr/dto/ObjectiveDto.java
@@ -12,11 +12,11 @@ public class ObjectiveDto {
     private Integer quarterNumber;
     private Integer quarterYear;
     private String description;
-    private Double progress;
+    private Long progress;
 
     public ObjectiveDto(Long id, String title, Long ownerId, String ownerFirstname, String ownerLastname, Long teamId,
             String teamName, Long quarterId, Integer quarterNumber, Integer quarterYear, String description,
-            Double progress) {
+            Long progress) {
         this.id = id;
         this.title = title;
         this.ownerId = ownerId;
@@ -119,11 +119,11 @@ public class ObjectiveDto {
         this.description = description;
     }
 
-    public Double getProgress() {
+    public Long getProgress() {
         return progress;
     }
 
-    public void setProgress(Double progress) {
+    public void setProgress(Long progress) {
         this.progress = progress;
     }
 }
